feat[fastproto]: update fastproto api.
1. update fastproto api;
2. fix bugs of crypto flow.

+++ src/main/java/org/indunet/fastproto/pipeline/AbstractFlow.java
@@ -43,5 +43,5 @@ public abstract class AbstractFlow<T> {
         this.next = null;
     }
 
-    public abstract int getFlowCode();
+    public abstract long getFlowCode();
 }
