+++ src/main/java/org/indunet/fastproto/pipeline/decode/VerifyChecksumFlow.java
@@ -31,7 +31,7 @@ import org.indunet.fastproto.pipeline.CodecContext;
  * @since 1.7.0
  */
 public class VerifyChecksumFlow extends AbstractFlow<CodecContext> {
-    public static final int FLOW_CODE = 0x0002;
+    public static final long FLOW_CODE = 0x0002;
 
     @Override
     public void process(CodecContext context) {
@@ -51,7 +51,7 @@ public class VerifyChecksumFlow extends AbstractFlow<CodecContext> {
     }
 
     @Override
-    public int getFlowCode() {
+    public long getFlowCode() {
         return FLOW_CODE;
     }
 }
