+++ src/main/java/org/indunet/fastproto/pipeline/encode/WriteProtocolVersionFlow.java
@@ -27,7 +27,7 @@ import org.indunet.fastproto.pipeline.CodecContext;
  * @since 1.7.0
  */
 public class WriteProtocolVersionFlow extends AbstractFlow<CodecContext> {
-    public static final int FLOW_CODE = 0x0400;
+    public static final long FLOW_CODE = 0x0400;
 
     @Override
     public void process(CodecContext context) {
@@ -37,7 +37,7 @@ public class WriteProtocolVersionFlow extends AbstractFlow<CodecContext> {
     }
 
     @Override
-    public int getFlowCode() {
+    public long getFlowCode() {
         return FLOW_CODE;
     }
 }
